chore: split stories in separate files
