+++ elements/timecontrol/stories/index.js
@@ -1,0 +1,7 @@
+// Global import of eox-elements in .storybook/preview.js!
+// Export different stories
+
+export { default as PrimaryStory } from "./primary"; // Primary story
+export { default as SliderStory } from "./slider"; // Slider story
+export { default as ProgrammaticTimeSelectionStory } from "./programmatic-time-selection"; // Programmatic time selection story
+export { default as DisabledPlayButtonStory } from "./disabled-play-button"; // Disabled play button story
